tech: Refactored audio context files.
