+++ src/audio/context/index.ts
@@ -1,3 +1,5 @@
 export * from './Context';
 export * from './Provider';
+export * from './use-context';
+export * from './use-provider';
