+++ src/audio/context/use-provider.ts
@@ -1,4 +1,4 @@
 import { useContext } from 'react';
-import { AudioContext } from '../context';
+import { AudioContext } from './Context';
 
 export const useAudio = () => useContext(AudioContext);
