+++ src/audio/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './use-audio';
 export * from './use-interval-storage';
 export * from './use-player';
 export * from './use-recorder';
