+++ src/audio/index.ts
@@ -1,3 +1,4 @@
+export * from './context';
 export * from './hooks';
 export * from './types';
 export * from './utils';
